Revert "소스트리 커밋 테스트"
This reverts commit e0cf24fca800fd20fe2b90f2cb334ca02678192b.

+++ src/main/java/org/zerock/mreview/repository/ReviewRepository.java
@@ -3,5 +3,5 @@ package org.zerock.mreview.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.zerock.mreview.entity.Review;
 
-public interface ReviewRepository extends JpaRepository<Review, Integer> {
+public interface ReviewRepository extends JpaRepository<Review, Long> {
 }
